Separated Swarm API, Swarm Social API and Social API
