+++ src/PostFeed.ts
@@ -2,7 +2,7 @@ import { Feed } from './models/Feed';
 import { PublicPost, Author } from './models/Post';
 import { ImageData } from './models/ImageData';
 import * as Swarm from './swarm/Swarm';
-import { uploadPost, uploadImage } from './PostUpload';
+import { uploadPost, uploadImage } from './swarm-social/swarmStorage';
 import { Debug } from './Debug';
 import { ModelHelper } from './models/ModelHelper';
